Add supervisor that is the process that starts the server
If the process that starts the program was passed to a user and
the user terminates, then all the program terminates. Being a supervisor
this process continues to work and the server continues too.

That way all the processes when are about to exit send a message to
supervisor with some data for him to know what to do next.